updated delyaed_job to use the capistrno gem instead of systemd like the cas app does

+++ lib/capistrano/tasks/delayed_job.rake
@@ -3,20 +3,26 @@
 namespace :delayed_job do
   task :restart do
     on roles(:job) do
-      execute :sudo, "bash -l -c 'systemctl stop delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service || echo ok'"
-      execute :sudo, "systemctl start delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service"
+      if ENV['DELAYED_JOB_SYSTEMD']=='true'
+        execute :sudo, "bash -l -c 'systemctl stop delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service || echo ok'"
+        execute :sudo, "systemctl start delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service"
+      end
     end
   end
 
   task :stop do
     on roles(:job) do
-      execute :sudo, "systemctl stop delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service"
+      if ENV['DELAYED_JOB_SYSTEMD']=='true'
+        execute :sudo, "systemctl stop delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service"
+      end
     end
   end
 
   task :start do
     on roles(:job) do
-      execute :sudo, "systemctl start delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service"
+      if ENV['DELAYED_JOB_SYSTEMD']=='true'
+        execute :sudo, "systemctl start delayed_job-#{fetch(:client)}-hmis-#{fetch(:rails_env)}.service"
+      end
     end
   end
 end
